requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
Flask-Login==0.6.3
Flask-Admin==2.0.2
Flask-Migrate==4.1.0
Flask-Bcrypt==1.0.1
Flask-CKEditor==1.0.0
Flask-CORS==6.0.1
Flask-Compress==1.23
Flask-Assets==2.1.0
Flask-Limiter==3.5.0
WTForms==3.2.1
SQLAlchemy==2.0.44
Werkzeug==3.1.3
Jinja2==3.1.6
Pillow==12.0.0
python-dotenv==1.2.1
python-slugify==8.0.4
gunicorn==23.0.0
psycopg2-binary==2.9.11
email-validator==2.3.0
requests==2.32.5
stripe==11.1.1
bleach==6.1.0
Flask-Mail
Flask-SocketIO
simple-websocket
aiortc
aiohttp
cryptography

